+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### Features
 * new API `terminal::animation()` and `terminal::delay()` [#683](https://github.com/jcubic/jquery.terminal/issues/683)
 * add images in less now have `terminal-less` class
-* add `history` option to `terminal::exec()` [#745](https://github.com/jcubic/jquery.terminal/issues/745)
+* add `execHistory` option to the terminal and `history` option to `terminal::exec()` [#745](https://github.com/jcubic/jquery.terminal/issues/745)
 ### Bugfix
 * fix number of rows when using different `--size` [#969](https://github.com/jcubic/jquery.terminal/issues/969)
 * fix hidden cursor when using command that change `--size` [#968](https://github.com/jcubic/jquery.terminal/issues/968)